Refactored Zones
Refactored zones into "Areas" given Destroy and Create need to be given areas similar to Zones but Zones are different entities from Destroy and Create Labor Order Areas. Also removed early resource backend for a different PR (this one is already getting big).
