+++ Assets/Scripts/World/Area.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: a74025754ac98374da4ba1c8b5797a52
+guid: c414f2386b84c3245b97d717239f2162
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
