--- Assets/Scripts/World/CropResource.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: e8bed5e145c81724196a002ec60b7b77
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
