--- Assets/Scripts/World/FarmZone.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 2d15cd3ea5dfe6746aada1c704330c46
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
