--- Assets/Scripts/World/Resource.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 447beb01267377e4fb8a9eceabdb373e
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
