--- Assets/Scripts/World/Value.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: ca10dbd7a0ce892468bdea2c48a14020
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
