--- Assets/Scripts/World/VisualZone.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 18c66d6923d10e542ae579c255fb50f2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
