+++ Assets/Scripts/World/Zone.cs
@@ -2,38 +2,18 @@ using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-public class Zone
+public class Zone : Area
 {
-    public enum ZoneType { NormalFoodFarm = 1, PremiumFoodFarm = 2, ClothFarm = 3, NormalMedicineFarm = 4, PremiumMedicineFarm = 5, Destroy = 6, Create = 7 };
+    public enum ZoneType { NormalFoodFarm = 1, PremiumFoodFarm = 2, ClothFarm = 3, NormalMedicineFarm = 4, PremiumMedicineFarm = 5 };
 
     protected ZoneType myType;
-
-    protected Vector2 topRight;
-    protected Vector2 bottomLeft;
-    protected List<Vector2> corners;
-
-    protected float height;
-    protected float width;
-    protected Vector2 middle;
-
     protected Colony myColony;
+    protected GameObject visualBox;
+    protected string zoneName;
 
-    public Zone(Vector2 tR, Vector2 bL, Colony colony, int type)
+    public Zone(Vector2 tR, Vector2 bL, Colony colony, int type) : base(tR, bL)
     {
         myColony = colony;
-
-        topRight = tR;
-        bottomLeft = bL;
-        List<Vector2> corners = new List<Vector2>();
-        corners.Add(topRight);                  //Top right
-        corners.Add(bottomLeft);                //Bottom left
-        corners.Add(new Vector2(bL.x, tR.y));   //Top left
-        corners.Add(new Vector2(tR.x, bL.y));   //Bottom right
-
-        height = topRight.y - bottomLeft.y;
-        width = topRight.x - bottomLeft.x;
-
-        middle = new Vector2(bottomLeft.x + (width/2), bottomLeft.y + (height/2));
 
         if(type < 7 && type > 1)
         {
@@ -43,35 +23,39 @@ public class Zone
         {
             myType = ZoneType.NormalFoodFarm;
         }
-    }
 
-    public Zone()
-    {
-        topRight = new Vector2();
-        bottomLeft = new Vector2();
-        corners.Add(topRight);                  //Top right
-        corners.Add(bottomLeft);                //Bottom left
-        corners.Add(new Vector2());             //Top left
-        corners.Add(new Vector2());             //Bottom right
-    }
+        visualBox = new GameObject("Zone" + myColony.GetNextZoneNumber());
+        zoneName = "Zone " + myColony.GetNextZoneNumber();
+        visualBox.transform.position = new Vector3(middle.x, middle.y, 0);
+        visualBox.transform.localScale = new Vector3(width, height, 1f);
+        visualBox.transform.SetParent(colony.gameObject.transform);
+        SpriteRenderer myRend = visualBox.AddComponent<SpriteRenderer>();
+        myRend.sprite = myColony.GetZoneSprite();
 
-    public List<Vector2> GetTwoCorners()
-    {
-        return new List<Vector2>{topRight, bottomLeft};
-    }
+        Color tmp = Color.yellow;
+        tmp.a = 0.35f;
+        myRend.color = tmp;
 
-    public List<Vector2> GetCorners()
-    {
-        return corners;
-    }
-
-    public Vector2 GetCenter()
-    {
-        return middle;
+        colony.AddZone(this);
     }
 
     public ZoneType GetZoneType()
     {
         return myType;
+    }
+
+    public string GetZoneName()
+    {
+        return zoneName;
+    }
+
+    public virtual void DeleteZone()
+    {
+        myColony.RemoveZone(this);
+    }
+
+    public GameObject GetVisualBox()
+    {
+        return visualBox;
     }
 }
